requirements.txt
dpkt==1.9.8
ipinfo==4.4.3
python-dotenv==0.20.0
